refactor: small fixes for CardComponent
