+++ AdyenCard/Components/Stored Card/StoredCardConfiguration.swift
@@ -1,0 +1,17 @@
+//
+// Copyright (c) 2021 Adyen N.V.
+//
+// This file is open source and available under the MIT license. See the LICENSE file for more info.
+//
+
+import Foundation
+
+/// Stored card configuration.
+public struct StoredCardConfiguration {
+
+    /// Indicates whether to show the security code field.
+    public var showsSecurityCodeField = true
+
+    /// :nodoc:
+    public init() { /* empty init */ }
+}
